Pick a reasonable default leelaz binary.
Now that Leela-Zero has the new GTP extensions that Lizzie requires,
it's no longer mandatory to compile a custom leelaz binary, which means
it's no longer a hard requirement for there to be a leelaz binary in
Lizzie's current working directory.

To make life easier for packagers (like me), when constructing the
default config.txt, Lizzie should examine the PATH and find the
system-wide leelaz. Of course, a local ./leelaz should still take
priority (even if `.` is not actually on your PATH).

This makes that the default behaviour. Works for me on a fresh install
with leela-zero installed in /usr/bin.