Remove all nulls (#402)
* Remove all nulls

Refactors the entire code base to remove all usages for null as a
values, and all null checks. When needed, data structures are updated to
use java.util.Optional, but in a lot of cases null are removed without
replacement.
Reimplement nextBranch and previousBranch
Minor graphical adjustment
Fix the Lizzie window not actually starting centered.
Remove the 0.98 factor on board size to remove the small region of
background above the board.
Fix an exception when creating a BufferedImage with width zero.